pyproject.toml
[tool.poetry]
name = "areyougreen2"
version = "0.1.0"
description = ""
authors = ["Jordan G <itsseraphii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.15.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-SocketIO = "^5.0.1"
SQLAlchemy = "^1.3.22"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
